chess_ai_commentary/backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
chess
stockfish
google-generativeai
TTS
torch
torchaudio
supabase
pydantic
chess.com
pygame